Rename `pj.h` to `rz_pj.h`
